# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_ms-kanban-server_internal_handlers_dto_request.CreateUserStoryStatusRequest:
    properties:
      color:
        type: string
      display_order:
        minimum: 0
        type: integer
      is_closed:
        type: boolean
      is_final:
        type: boolean
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - color
    - name
    type: object
...
  github_com_ms-kanban-server_internal_handlers_dto_request.UpdateUserStoryStatusRequest:
    properties:
      color:
        type: string
      display_order:
        minimum: 0
        type: integer
      is_closed:
        type: boolean
      is_final:
        type: boolean
      name:
        maxLength: 50
        minLength: 1
        type: string
    type: object
...
  github_com_ms-kanban-server_internal_handlers_dto_response.UserStoryStatusResponse:
    properties:
      color:
        type: string
      display_order:
        type: integer
      id:
        type: string
      is_closed:
        type: boolean
      is_default:
        type: boolean
      is_final:
        type: boolean
      name:
        type: string
      project_id:
        type: string
    type: object
...
  github_com_ms-kanban-server_internal_pkg_response.Error:
    properties:
      code:
        $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorCode'
      message:
        type: string
      status_code:
        type: integer
    type: object
  github_com_ms-kanban-server_internal_pkg_response.ErrorCode:
    enum:
    - BAD_REQUEST
    - INVALID_STATUS_TRANSITION
    - METHOD_NOT_ALLOWED
    - VALIDATION_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - RESOURCE_NOT_FOUND
    - CONFLICT
    - GONE
    - TASK_PERMANENTLY_DELETED
    - BUSINESS_RULE_VIOLATION
    - RATE_LIMIT_EXCEEDED
    - INTERNAL_SERVER_ERROR
    - SERVICE_UNAVAILABLE
    - GATEWAY_TIMEOUT
    type: string
    x-enum-varnames:
    - ErrBadRequest
    - ErrInvalidStatusTransition
    - ErrMethodNotAllowed
    - ErrValidation
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrConflict
    - ErrGone
    - ErrTaskPermanentlyDeleted
    - ErrBusinessRule
    - ErrRateLimitExceeded
    - ErrInternalServerError
    - ErrServiceUnavailable
    - ErrGatewayTimeout
  github_com_ms-kanban-server_internal_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.Error'
      success:
        type: boolean
    type: object
  github_com_ms-kanban-server_internal_pkg_response.Pagination:
    properties:
      has_next:
        type: boolean
      has_previous:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_ms-kanban-server_internal_pkg_response.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.Pagination'
      status_code:
        type: integer
      success:
        type: boolean
    type: object
info:
  contact:
    email: olivergot26@gmail.com
    name: PMT Team
  description: PMT Backend API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: PMT API
  version: "1.0"
paths:
...
  /projects/{project_id}/user-story-statuses:
    get:
      description: Get all User Story statuses for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.UserStoryStatusResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User Story statuses
      tags:
      - User Story Statuses
    post:
      consumes:
      - application/json
      description: Create a new User Story custom status for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Create User Story Status Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_request.CreateUserStoryStatusRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.UserStoryStatusResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a User Story status
      tags:
      - User Story Statuses
  /projects/{project_id}/user-story-statuses/{status_id}:
    delete:
      description: Delete an existing User Story custom status
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Status ID
        in: path
        name: status_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a User Story status
      tags:
      - User Story Statuses
    patch:
      consumes:
      - application/json
      description: Update an existing User Story custom status
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Status ID
        in: path
        name: status_id
        required: true
        type: string
      - description: Update User Story Status Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_request.UpdateUserStoryStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.UserStoryStatusResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a User Story status
      tags:
      - User Story Statuses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"